+++ go/ai/ai_service/provider/deepgram/text_to_speech.go
@@ -154,10 +154,10 @@ func (c *Client) TextToSpeechStream(
 	modelUsage := &aipb.ModelUsage{
 		Model: request.Model,
 		InputCharacter: &aipb.ResourceConsumption{
-			Quantity: int32(billableCharacterCount),
+			Quantity: int64(billableCharacterCount),
 		},
 		OutputSecond: &aipb.ResourceConsumption{
-			Quantity: int32(totalDuration.Seconds()),
+			Quantity: int64(totalDuration.Seconds()),
 		},
 	}
 	if err := srv.Send(&aiservicepb.TextToSpeechStreamResponse{
